+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning].
 
 ## [Unreleased]
 
+### Added
+
+- Support for `InertiaRails.scroll` attributes. ([@skryukov])
+
 ## [0.1.1] - 2025-11-05
 
 ### Added